SINGA-217 build python package with setup.py
CudnnRNN requires cudnn version later than 5.05, the previous check
based on CUDNN_VERSION_MAJOR and CUDNN_VERSION_PATCH does not work for
version 5.13. Therefore, we replaced the check to CUDNN_VERSION >= 5005.
CUDNN_VERSION is defined in cudnn.h.

cmake defines CUDNN_VERSION_SWIG which is used by model_layer.i for cudnn
rnn.

Remvoe the hardcoded sys imports for python files. Now users need to
install the python module to run python examples and tests.

Tested installation in python virtual enviornment and conda
virtural enviorment:

Inside build/python
```
pip install -e .
```

+++ src/io/csv_decoder.cc
@@ -33,7 +33,7 @@ std::vector<Tensor> CSVDecoder::Decode(std::string value) {
     ss >> l;
   std::string str;
   float d[kMaxCSVBufSize];
-  size_t size = 0;
+  int size = 0;
   while (std::getline(ss, str, ',')) {
     float temp;
     if (std::stringstream(str) >> temp) {
@@ -42,7 +42,7 @@ std::vector<Tensor> CSVDecoder::Decode(std::string value) {
     }
   }
 
-  Tensor data(Shape {size}, kFloat32);
+  Tensor data(Shape {static_cast<size_t>(size)}, kFloat32);
   data.CopyDataFromHostPtr(d, size);
   output.push_back(data);
   if (has_label_ == true) {
